[core] POC around material-ui-naked
