+++ packages/material-ui-naked/src/Dialog/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './Dialog';
+export * from './Dialog';
