+++ packages/material-ui-naked/src/Dialog/index.js
@@ -1,0 +1,2 @@
+export { default } from './Dialog';
+export * from './Dialog';
