+++ packages/material-ui-naked/src/index.js
@@ -1,0 +1,1 @@
+export * from './Dialog';
